Swipeable to ReanimatedSwipeable, TouchableWithoutFeedback to Pressable and folder name changes
